code/backend_auth/requirements.txt
fastapi
uvicorn
pydantic
python-jose
bcrypt
pymongo
python-dotenv
passlib
python-multipart